% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dictionaries.R
\docType{data}
\name{beaufort_scale_dictionary}
\alias{beaufort_scale_dictionary}
\title{Beaufort scale data dictionary}
\format{
A tibble with 4 rows and 3 variables:
\describe{
\item{column_name}{(\code{character}) Name of the column in \link{beaufort_scale}.}
\item{class}{(\code{character}) Class label for the column, following \code{vctrs}
type conventions.}
\item{description}{(\code{character}) Human-readable description of the
column's contents.}
}
}
\usage{
beaufort_scale_dictionary
}
\description{
Column names, class labels, and descriptions for each variable in
\link{beaufort_scale}.
}
\examples{
beaufort_scale_dictionary
}
\seealso{
\link{beaufort_scale}, the dataset this dictionary describes. The script
used to prepare this dataset:
\url{https://github.com/jonthegeek/seabirddata/blob/main/data-raw/dictionaries.R}
}
\keyword{datasets}
