% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dictionaries.R
\docType{data}
\name{birds_dictionary}
\alias{birds_dictionary}
\title{Bird observation records data dictionary}
\format{
A tibble with 26 rows and 3 variables:
\describe{
\item{column_name}{(\code{character}) Name of the column in \link{birds}.}
\item{class}{(\code{character}) Class label for the column, following \code{vctrs}
type conventions.}
\item{description}{(\code{character}) Human-readable description of the
column's contents.}
}
}
\usage{
birds_dictionary
}
\description{
Column names, class labels, and descriptions for each variable in \link{birds}.
}
\examples{
birds_dictionary
}
\seealso{
\link{birds}, the dataset this dictionary describes. The script used to
prepare this dataset:
\url{https://github.com/jonthegeek/seabirddata/blob/main/data-raw/dictionaries.R}
}
\keyword{datasets}
