% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dictionaries.R
\name{.class_friendly}
\alias{.class_friendly}
\title{Describe the class of an object for a data dictionary}
\usage{
.class_friendly(x)
}
\arguments{
\item{x}{(\code{any}) The object to describe.}
}
\value{
(\code{character(1)}) A descriptive class label for the object.
}
\description{
Determine a human-readable class label for an object.
}
\keyword{internal}
