% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seabirddata-package.R
\docType{package}
\name{seabirddata-package}
\alias{seabirddata}
\alias{seabirddata-package}
\title{seabirddata: What the Package Does}
\description{
The goal of seabirddata is to...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/seabirddata}
  \item \url{https://jonthegeek.github.io/seabirddata/}
  \item Report bugs at \url{https://github.com/jonthegeek/seabirddata/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
