% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dictionaries.R
\docType{data}
\name{ships_dictionary}
\alias{ships_dictionary}
\title{Ship observation records data dictionary}
\format{
A tibble with 21 rows and 3 variables:
\describe{
\item{column_name}{(\code{character}) Name of the column in \link{ships}.}
\item{class}{(\code{character}) Class label for the column, following \code{vctrs}
type conventions.}
\item{description}{(\code{character}) Human-readable description of the
column's contents.}
}
}
\usage{
ships_dictionary
}
\description{
Column names, class labels, and descriptions for each variable in \link{ships}.
}
\examples{
ships_dictionary
}
\seealso{
\link{ships}, the dataset this dictionary describes. The script used to
prepare this dataset:
\url{https://github.com/jonthegeek/seabirddata/blob/main/data-raw/dictionaries.R}
}
\keyword{datasets}
