% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seabirddata-package.R
\docType{package}
\name{seabirddata-package}
\alias{seabirddata}
\alias{seabirddata-package}
\title{seabirddata: Data of At-Sea Observations of Seabirds}
\description{
This is a dataset of at-sea observations of seabirds dating from 1969 to 1990.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/seabirddata}
  \item \url{https://jonthegeek.github.io/seabirddata/}
  \item Report bugs at \url{https://github.com/jonthegeek/seabirddata/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Other contributors:
\itemize{
  \item Museum of New Zealand Te Papa Tongarewa (This package uses data sourced from Te Papa and licensed by Te Papa for re-use under the Creative Commons BY 4.0 International licence) [copyright holder]
}

}
\keyword{internal}
